requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
faster-whisper==1.0.3
python-multipart==0.0.9
jinja2==3.1.6
pywebview==4.4.1
av==12.2.0
openai>=1.40.0,<2
httpx<0.27
huggingface_hub>=0.20.3
tqdm>=4.66
pyinstaller==6.15.0